refactor(activerecord): fold find-normalization helpers into finder-methods
Rails keeps the shared id-normalization + RecordNotFound helpers
(`find_with_ids`, `raise_record_not_found_exception!`) as private
methods on `ActiveRecord::FinderMethods`. Folding our pragmatic
`find-normalization.ts` extract back into `finder-methods.ts`
matches that layout — CollectionProxy#find imports the helpers from
the same module the SQL path lives in.

Also renames find-normalization.test.ts to finder-methods.test.ts to
align with the module it tests.
