+++ packages/activerecord/src/relation/finder-methods.ts
@@ -10,7 +10,183 @@
 
 import { RecordNotFound, SoleRecordExceeded } from "../errors.js";
 import { RecordInvalid } from "../validations.js";
-import { normalizeFindArgs, raiseNotFoundAll, raiseNotFoundSingle } from "./find-normalization.js";
+
+// ---------------------------------------------------------------------------
+// Shared id-normalization + not-found helpers.
+//
+// Both `Relation.performFind` (SQL path — below) and
+// `CollectionProxy#find` (in-memory association path — associations/
+// collection-proxy.ts) accept the same polymorphic arg set and must
+// produce identical `RecordNotFound` messages and `.id` payloads.
+// Centralizing the normalization + raise helpers here prevents drift.
+//
+// Simple-PK flattening uses `flat(Infinity)`, matching Ruby's
+// `Array#flatten` default — Rails contract (`Post.find([[1, [2]]])`
+// works as `Post.find(1, 2)`).
+//
+// Mirrors: ActiveRecord::FinderMethods' private
+// `find_with_ids` / `raise_record_not_found_exception!` helpers.
+// ---------------------------------------------------------------------------
+
+export interface NormalizedFindIds {
+  /**
+   * Canonical id list for the lookup backend:
+   *   - simple PK → flat scalar ids (never arrays).
+   *   - composite PK → array of tuples (always `unknown[][]`).
+   */
+  readonly ids: unknown[];
+
+  /**
+   * `true` when the caller provided a list-form (variadic ≥2, a
+   * single array arg, or composite list-of-tuples) and therefore
+   * wants `T[]` back. `false` for the single-id / single-tuple case.
+   */
+  readonly wantArray: boolean;
+
+  /**
+   * For composite PKs: the tuple list (same shape as `ids`). For
+   * simple PKs: `null`. Used to format error messages + payload
+   * exactly like Rails (`String(tuples)` vs `flatIds.join(", ")`).
+   */
+  readonly tuples: unknown[][] | null;
+}
+
+/**
+ * Normalize the varargs of a `.find(...)` call into the canonical
+ * `NormalizedFindIds` shape.
+ *
+ * Raises `RecordNotFound` for the deterministic input errors:
+ *   - zero-arg call                     → "empty list of ids", `id: []`
+ *   - explicit `find([])`               → same
+ *   - composite PK + scalar or wrong-arity tuple →
+ *     "`<Model>: composite primary key requires a <N>-element array, got <id>`"
+ *
+ * Does NOT do the actual lookup or the "couldn't find all" aggregate
+ * error — that stays at the call site (SQL vs in-memory each have
+ * their own count-comparison logic).
+ */
+export function normalizeFindArgs(
+  modelName: string,
+  pk: string | string[],
+  args: unknown[],
+): NormalizedFindIds {
+  const composite = Array.isArray(pk);
+
+  if (args.length === 0) {
+    throw new RecordNotFound(
+      `Couldn't find ${modelName} with an empty list of ids`,
+      modelName,
+      String(pk),
+      [],
+    );
+  }
+
+  const [first, ...rest] = args;
+  let ids: unknown[];
+  let wantArray: boolean;
+
+  if (rest.length > 0) {
+    if (composite) {
+      if (args.every((x) => !Array.isArray(x))) {
+        // All-scalar collapses to one tuple id — Rails treats
+        // `find(1, 42)` on a 2-arity PK as `[1, 42]`. Arity mismatch
+        // raises below with the whole tuple in the message.
+        ids = [args];
+        wantArray = false;
+      } else {
+        ids = args;
+        wantArray = true;
+      }
+    } else {
+      // Simple PK: flatten so mixed inputs like `find([1, 2], 3)`
+      // canonicalize to `[1, 2, 3]`.
+      ids = (args as unknown[]).flat(Infinity);
+      wantArray = true;
+    }
+  } else if (Array.isArray(first)) {
+    if (composite) {
+      if (first.length === 0) {
+        ids = first;
+        wantArray = true;
+      } else if (first.every((x) => !Array.isArray(x))) {
+        ids = [first];
+        wantArray = false;
+      } else {
+        ids = first;
+        wantArray = true;
+      }
+    } else {
+      // Simple PK: recursive flatten so `find([[1, 2]])` behaves like
+      // `find([1, 2])`, matching Rails' `Array#flatten`.
+      ids = (first as unknown[]).flat(Infinity);
+      wantArray = true;
+    }
+  } else {
+    ids = [first];
+    wantArray = false;
+  }
+
+  if (ids.length === 0) {
+    throw new RecordNotFound(
+      `Couldn't find ${modelName} with an empty list of ids`,
+      modelName,
+      String(pk),
+      [],
+    );
+  }
+
+  if (composite) {
+    const pkArity = (pk as string[]).length;
+    for (const id of ids) {
+      if (!Array.isArray(id) || id.length !== pkArity) {
+        throw new RecordNotFound(
+          `${modelName}: composite primary key requires a ${pkArity}-element array, got ${String(id)}`,
+          modelName,
+          String(pk),
+          id,
+        );
+      }
+    }
+    return { ids, wantArray, tuples: ids as unknown[][] };
+  }
+
+  return { ids, wantArray, tuples: null };
+}
+
+/**
+ * Raise the aggregate "couldn't find all" error, matching
+ * `Relation.performFind`'s message shape for the caller's PK kind:
+ *   - simple PK  → `flatIds.join(", ")`, payload = flatIds.
+ *   - composite  → `String(tuples)`    , payload = tuples[][].
+ */
+export function raiseNotFoundAll(
+  modelName: string,
+  pk: string | string[],
+  normalized: NormalizedFindIds,
+): never {
+  const { ids, tuples } = normalized;
+  const messageIds = tuples ? String(tuples) : (ids as unknown[]).join(", ");
+  const payload = tuples ?? ids;
+  throw new RecordNotFound(
+    `Couldn't find all ${modelName} with '${String(pk)}': (${messageIds})`,
+    modelName,
+    String(pk),
+    payload,
+  );
+}
+
+/**
+ * Raise the single-id not-found error for a simple PK.
+ * Matches `Relation.performFind`'s `"with 'pk'=<id>"` message.
+ */
+export function raiseNotFoundSingle(modelName: string, pk: string, id: unknown): never {
+  throw new RecordNotFound(
+    `Couldn't find ${modelName} with '${pk}'=${String(id)}`,
+    modelName,
+    pk,
+    id,
+  );
+}
 
 interface FinderRelation {
   _modelClass: {
